Conflicts:
	Yeshivat Torat Shraga/Code/Models/SearchModel.swift
	Yeshivat Torat Shraga/Code/Structure/FirebaseConnection.swift
	Yeshivat Torat Shraga/Code/Views/RabbiCardView.swift
	Yeshivat Torat Shraga/Code/functions/src/index.ts

+++ Yeshivat Torat Shraga/Code/Structure/Favorites.swift
@@ -10,22 +10,50 @@ import UIKit
 import SwiftUI
 import CoreData
 
-class Favorites {
-    static let delegate = (UIApplication.shared.delegate as! AppDelegate)
+class Favorites: ObservableObject {
+    static var shared = Favorites()
+    let delegate = (UIApplication.shared.delegate as! AppDelegate)
     typealias FavoritesTuple = (videos: [Video]?, audios: [Audio]?, people: [DetailedRabbi]?)
     
-    private static var favorites: FavoritesTuple?// = loadFavorites()
+    init() {
+        favoriteIDs = getfavoriteIDs()
+    }
+    
+    @Published var favoriteIDs: [FirestoreID]?
+    @Published var favorites: FavoritesTuple?
+    
+     func getfavoriteIDs() -> [FirestoreID] {
+        var IDs: [FirestoreID] = []
+         if let favorites = self.loadFavorites() {
+            if let videos = favorites.videos {
+                for video in videos {
+                    IDs.append(video.firestoreID)
+                }
+            }
+            if let audios = favorites.audios {
+                for audio in audios {
+                    IDs.append(audio.firestoreID)
+                }
+            }
+            if let people = favorites.people {
+                for person in people {
+                    IDs.append(person.firestoreID)
+                }
+            }
+        }
+        return IDs
+    }
     
     /// Retreives the most updated favorites tuple for the device.
-    static func getFavorites(completion: @escaping ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)) {
+    func getFavorites(completion: @escaping ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)) {
         if let favorites = favorites {
             completion(favorites, nil)
         } else {
-            loadFavorites(completion: completion)
+            self.loadFavorites(completion: completion)
         }
     }
     
-    static func clearFavorites() {
+    func clearFavorites() {
         let entities = [CDVideo.entity(), CDAudio.entity(), CDPerson.entity()]
         for entity in entities {
             let fetchRequest = NSFetchRequest<NSFetchRequestResult>(entityName: entity.name!)
@@ -38,7 +66,7 @@ class Favorites {
         }
     }
     
-    static func save(_ rabbiToSave: DetailedRabbi, completion: ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)? = nil) {
+    func save(_ rabbiToSave: DetailedRabbi, completion: ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)? = nil) {
         DispatchQueue.global(qos: .background).async {
             let group = DispatchGroup()
             
@@ -52,7 +80,7 @@ class Favorites {
                 DispatchQueue.main.async {
                     guard let data = profilePicture.asUIImage().jpegData(compressionQuality: 1.0) else {
                         DispatchQueue.main.async {
-                            loadFavorites(completion: completion)
+                            self.loadFavorites(completion: completion)
                         }
                         return
                     }
@@ -62,7 +90,7 @@ class Favorites {
             } else if let profilePictureURL = profilePictureURL {
                 guard let data = try? Data(contentsOf: profilePictureURL) else {
                     DispatchQueue.main.async {
-                        loadFavorites(completion: completion)
+                        self.loadFavorites(completion: completion)
                     }
                     return
                 }
@@ -72,7 +100,7 @@ class Favorites {
                 group.leave()
             }
             
-            let managedContext = delegate.persistentContainer.viewContext
+            let managedContext = self.delegate.persistentContainer.viewContext
             
             var cdAuthor: CDPerson
             cdAuthor = CDPerson(context: managedContext)
@@ -87,17 +115,17 @@ class Favorites {
                 DispatchQueue.main.async {
                     do {
                         try managedContext.save()
-                        loadFavorites(completion: completion)
+                        self.loadFavorites(completion: completion)
                     } catch let error as NSError {
                         print("Failed to save: \(error), \(error.userInfo)")
-                        loadFavorites(completion: completion)
+                        self.loadFavorites(completion: completion)
                     }
                 }
             }
         }
     }
     
-    static func save(_ videoToSave: Video, completion: ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)? = nil) {
+    func save(_ videoToSave: Video, completion: ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)? = nil) {
         DispatchQueue.global(qos: .background).async {
             let group = DispatchGroup()
             
@@ -111,7 +139,7 @@ class Favorites {
                 DispatchQueue.main.async {
                     guard let data = authorProfilePicture.asUIImage().jpegData(compressionQuality: 1.0) else {
                         DispatchQueue.main.async {
-                            loadFavorites(completion: completion)
+                            self.loadFavorites(completion: completion)
                         }
                         return
                     }
@@ -121,7 +149,7 @@ class Favorites {
             } else if let authorProfilePictureURL = authorProfilePictureURL {
                 guard let data = try? Data(contentsOf: authorProfilePictureURL) else {
                     DispatchQueue.main.async {
-                        loadFavorites(completion: completion)
+                        self.loadFavorites(completion: completion)
                     }
                     return
                 }
@@ -141,7 +169,7 @@ class Favorites {
                 DispatchQueue.main.async {
                     guard let data = thumbnail.asUIImage().jpegData(compressionQuality: 1.0) else {
                         DispatchQueue.main.async {
-                            loadFavorites(completion: completion)
+                            self.loadFavorites(completion: completion)
                         }
                         return
                     }
@@ -151,7 +179,7 @@ class Favorites {
             } else if let thumbnailURL = thumbnailURL {
                 guard let data = try? Data(contentsOf: thumbnailURL) else {
                     DispatchQueue.main.async {
-                        loadFavorites(completion: completion)
+                        self.loadFavorites(completion: completion)
                     }
                     return
                 }
@@ -164,12 +192,12 @@ class Favorites {
             
             guard let duration = videoToSave.duration else {
                 DispatchQueue.main.async {
-                    loadFavorites(completion: completion)
+                    self.loadFavorites(completion: completion)
                 }
                 return
             }
             
-            let managedContext = delegate.persistentContainer.viewContext
+            let managedContext = self.delegate.persistentContainer.viewContext
             
             let entity = CDVideo.entity()
             
@@ -180,6 +208,7 @@ class Favorites {
             cdVideo.fileID = videoToSave.fileID
             cdVideo.title = videoToSave.title
             cdVideo.body = videoToSave.description
+//            cdVideo.favoritedAt = videoToSave.favoritedAt
             //            MARK: NOT SAVING TAGS
             //            cdAudio.tags = audioToSave.tags
             cdVideo.uploadDate = videoToSave.date
@@ -203,17 +232,17 @@ class Favorites {
                 DispatchQueue.main.async {
                     do {
                         try managedContext.save()
-                        loadFavorites(completion: completion)
+                        self.loadFavorites(completion: completion)
                     } catch let error as NSError {
                         print("Failed to save: \(error), \(error.userInfo)")
-                        loadFavorites(completion: completion)
+                        self.loadFavorites(completion: completion)
                     }
                 }
             }
         }
     }
     
-    static func save(_ audioToSave: Audio, completion: ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)? = nil) {
+    func save(_ audioToSave: Audio, completion: ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)? = nil) {
         DispatchQueue.global(qos: .background).async {
             let group = DispatchGroup()
             
@@ -227,7 +256,7 @@ class Favorites {
                 DispatchQueue.main.async {
                     guard let data = authorProfilePicture.asUIImage().jpegData(compressionQuality: 1.0) else {
                         DispatchQueue.main.async {
-                            loadFavorites(completion: completion)
+                            self.loadFavorites(completion: completion)
                         }
                         return
                     }
@@ -237,7 +266,7 @@ class Favorites {
             } else if let authorProfilePictureURL = authorProfilePictureURL {
                 guard let data = try? Data(contentsOf: authorProfilePictureURL) else {
                     DispatchQueue.main.async {
-                        loadFavorites(completion: completion)
+                        self.loadFavorites(completion: completion)
                     }
                     return
                 }
@@ -249,12 +278,12 @@ class Favorites {
             
             guard let duration = audioToSave.duration else {
                 DispatchQueue.main.async {
-                    loadFavorites(completion: completion)
+                    self.loadFavorites(completion: completion)
                 }
                 return
             }
             
-            let managedContext = delegate.persistentContainer.viewContext
+            let managedContext = self.delegate.persistentContainer.viewContext
             
             let entity = CDAudio.entity()
             
@@ -265,6 +294,7 @@ class Favorites {
             cdAudio.fileID = audioToSave.fileID
             cdAudio.title = audioToSave.title
             cdAudio.body = audioToSave.description
+//            cdAudio.favoritedAt = audioToSave.favoritedAt
             //            MARK: NOT SAVING TAGS
             //            cdAudio.tags = audioToSave.tags
             cdAudio.uploadDate = audioToSave.date
@@ -287,93 +317,93 @@ class Favorites {
                 DispatchQueue.main.async {
                     do {
                         try managedContext.save()
-                        loadFavorites(completion: completion)
+                        self.loadFavorites(completion: completion)
                     } catch let error as NSError {
                         print("Failed to save: \(error), \(error.userInfo)")
-                        loadFavorites(completion: completion)
+                        self.loadFavorites(completion: completion)
                     }
                 }
             }
         }
     }
     
-    static func delete(_ rabbiToDelete: DetailedRabbi, completion: ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)? = nil) {
+    func delete(_ rabbiToDelete: DetailedRabbi, completion: ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)? = nil) {
         let managedContext = delegate.persistentContainer.viewContext
         let fetchRequest = CDPerson.fetchRequest()
         
         do {
             let result = try managedContext.fetch(fetchRequest)
             
-                if let match = result.first(where: { r in
-                    r.firestoreID == rabbiToDelete.firestoreID
-                }) {
-                    managedContext.delete(match)
-                    
-                    do {
-                        try managedContext.save()
-                        loadFavorites(completion: completion)
-                    } catch {
-                        print("Failed to delete: \(error)")
-                        loadFavorites(completion: completion)
-                    }
+            if let match = result.first(where: { r in
+                r.firestoreID == rabbiToDelete.firestoreID
+            }) {
+                managedContext.delete(match)
+                
+                do {
+                    try managedContext.save()
+                    loadFavorites(completion: completion)
+                } catch {
+                    print("Failed to delete: \(error)")
+                    loadFavorites(completion: completion)
                 }
+            }
         } catch {
             print("Failed to delete: \(error)")
         }
     }
     
-    static func delete(_ videoToDelete: Video, completion: ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)? = nil) {
+    internal func delete(_ videoToDelete: Video, completion: ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)? = nil) {
         let managedContext = delegate.persistentContainer.viewContext
         let fetchRequest = CDVideo.fetchRequest()
         
         do {
             let result = try managedContext.fetch(fetchRequest)
             
-                if let match = result.first(where: { v in
-                    v.firestoreID == videoToDelete.firestoreID
-                }) {
-                    managedContext.delete(match)
-                    
-                    do {
-                        try managedContext.save()
-                        loadFavorites(completion: completion)
-                    } catch {
-                        print("Failed to delete: \(error)")
-                        loadFavorites(completion: completion)
-                    }
+            if let match = result.first(where: { v in
+                v.firestoreID == videoToDelete.firestoreID
+            }) {
+                managedContext.delete(match)
+                
+                do {
+                    try managedContext.save()
+                    loadFavorites(completion: completion)
+                } catch {
+                    print("Failed to delete: \(error)")
+                    loadFavorites(completion: completion)
                 }
+            }
         } catch {
             print("Failed to delete: \(error)")
         }
     }
     
-    static func delete(_ audioToDelete: Audio, completion: ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)? = nil) {
+    func delete(_ audioToDelete: Audio, completion: ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)? = nil) {
         let managedContext = delegate.persistentContainer.viewContext
         let fetchRequest = CDAudio.fetchRequest()
         
         do {
             let result = try managedContext.fetch(fetchRequest)
             
-                if let match = result.first(where: { a in
-                    a.firestoreID == audioToDelete.firestoreID
-                }) {
-                    managedContext.delete(match)
-                    
-                    do {
-                        try managedContext.save()
-                        loadFavorites(completion: completion)
-                    } catch {
-                        print("Failed to delete: \(error)")
-                        loadFavorites(completion: completion)
-                    }
+            if let match = result.first(where: { a in
+                a.firestoreID == audioToDelete.firestoreID
+            }) {
+                managedContext.delete(match)
+                
+                do {
+                    try managedContext.save()
+                    loadFavorites(completion: completion)
+                } catch {
+                    print("Failed to delete: \(error)")
+                    loadFavorites(completion: completion)
                 }
+            }
         } catch {
             print("Failed to delete: \(error)")
         }
     }
     
-    private static func loadFavorites(completion: ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)? = nil) {
-        let managedContext = Favorites.delegate.persistentContainer.viewContext
+    func loadFavorites(completion: ((_ favorites: FavoritesTuple?, _ error: Error?) -> Void)? = nil) {
+        let managedContext = self.delegate.persistentContainer.viewContext
         
         let group = DispatchGroup()
         let queue = DispatchQueue(label: "favorites_queue", attributes: .concurrent)
@@ -444,12 +474,13 @@ class Favorites {
         group.notify(queue: .main) {
             let favorites = (videos: favoriteVideos, audios: favoriteAudios, people: favoritePeople)
             self.favorites = favorites
+            self.favoriteIDs = self.getfavoriteIDs()
             completion?(favorites, nil)
         }
     }
     
-    private static func loadFavorites() -> FavoritesTuple? {
-        let managedContext = Favorites.delegate.persistentContainer.viewContext
+    func loadFavorites() -> FavoritesTuple? {
+        let managedContext = self.delegate.persistentContainer.viewContext
         
         var favoritePeople: [DetailedRabbi]? = nil
         let peopleFetchRequest = CDPerson.fetchRequest()
